Update Burner.java
Replaced updateTemperature() with new version in line with new understanding of desired behavior. Old version still present but commented out. 

Note that a slight change to the declaration of the Temperature enum was needed for this to work. (must be "public enum Temperature {COOL, WARM, HOT, BLAZING};" with values listed in that exact order)